+++ pkg/website/urls.go
@@ -9,4 +9,5 @@ const (
 	BugReportURL              = "https://flamenco.blender.org/get-involved"
 	ShamanRequirementsURL     = "https://flamenco.blender.org/usage/shared-storage/shaman/#requirements"
 	WorkerConfigURL           = "https://flamenco.blender.org/usage/worker-configuration/"
+	OOMScoreAdjURL            = WorkerConfigURL
 )